+++ pet_shop/cliente/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class ClienteConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'cliente'
