+++ pet_shop/pet_shop/settings.py
@@ -45,6 +45,9 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'publica',
     'administracion',
+    'blog',
+    'cliente',
+    'carrito',
 ]
 
 MIDDLEWARE = [
