+++ pet_shop/publica/urls.py
@@ -9,10 +9,10 @@ urlpatterns = [
     path('blog/', views.blog, name='blog'),
     path('contactanos/', views.contactanos, name='contactanos'),
     path('carrito', views.carrito, name='carrito'),
-    path('checkout', views.checkout, name='checkout'),
+    
     path('favoritos', views.favoritos, name='favoritos'),
     path('product_single', views.producto_simple, name='producto_simple'),
-    path('registro', views.registro, name='registro'),
+    
     path('product_single', views.producto_simple, name='producto_simple'),
     path('logput/', views.exit, name='exit'),
 ]
