+++ pet_shop/templates/cliente/favoritos.html
@@ -1,8 +1,8 @@
-{% extends './base.html' %}
+{% extends './publica/base.html' %}
 {% load static %}
 
 {% block contenido %}
-{% include './banner.html' %}
+{% include './publica/banner.html' %}
 
     <section class="ftco-section ftco-cart">
 			<div class="container">
