+++ pet_shop/templates/cliente/pedidos_pendientes.html
@@ -1,0 +1,1 @@
+<h1>PEDIDOS PENDIENTES CLIENTE</h1>
