+++ pet_shop/templates/cliente/perfil.html
@@ -1,0 +1,1 @@
+<h1>Perfil Cliente</h1>
