+++ pet_shop/templates/cliente/registro.html
@@ -1,9 +1,9 @@
-{% extends './base.html' %}
+{% extends './publica/base.html' %}
 {% load static %}
 {% block css %}{% endblock css %}
 {% block contenido %}
 {% comment %} banner {% endcomment %}
-{% include './banner.html' %}
+{% include './publica/banner.html' %}
 
 
 <section class="ftco-section">
