+++ pet_shop/templates/publica/checkout.html
@@ -205,7 +205,6 @@
 
 		var quantitiy=0;
 		   $('.quantity-right-plus').click(function(e){
-		        
 		        // Stop acting like a button
 		        e.preventDefault();
 		        // Get the field name
@@ -219,7 +218,6 @@
 		            // Increment
 		        
 		    });
-
 		     $('.quantity-left-minus').click(function(e){
 		        // Stop acting like a button
 		        e.preventDefault();
